// src/types/WorkStoreType.ts
export interface ImgInfo {
    // Natural Width
    nw: number;
    // Natural Height
    nh: number;

    w: number;
    h: number;
    cw: number;
    ch: number;
}

export interface WorkSize {
    width: number,
    height: number,
}

export type SupportImageSource = HTMLImageElement | HTMLCanvasElement;